Rename pcsdk module to commands
