+++ commands/commands_test.go
